pyproject.toml
[tool.poetry]
name = "noohub"
version = "1.0.0"
description = "A global registry for Noofiles"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"
repository = "https://github.com/vcokltfre/noohub"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.19.2"
uvicorn = "^0.17.1"
ormar = "^0.10.24"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
taskipy = "^1.9.0"
alembic = "^1.7.6"
fastapi = "^0.73.0"
aioredis = "^2.0.1"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pre-commit = "^2.17.0"

[tool.taskipy.tasks]
dev = "uvicorn src.main:app --port 8080 --reload"
start = "uvicorn src.main:app --port 8080 --host 0.0.0.0"
lint = "black . && isort --profile black ."
pre-commit = "pre-commit install"
upgrade = "alembic upgrade head"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
